fix: keep native tab drag metadata process-private
