+++ Packages/macOS/CmuxSidebar/Sources/CmuxSidebar/Drag/SidebarWorkspaceDragParticipantReference.swift
@@ -1,0 +1,4 @@
+/// Weak reference to one window-local presentation of a workspace drag.
+struct SidebarWorkspaceDragParticipantReference {
+    weak var state: SidebarDragState?
+}
